+++ src/samplers/smc.jl
@@ -1,5 +1,5 @@
 immutable SMC <: InferenceAlgorithm
-  n_particles :: Int
+  n_particles :: Int64
   resampler :: Function
   resampler_threshold :: Float64
   SMC(n) = new(n, resampleSystematic, 0.5)
